Refactor: Extract PDF image parsing and text extraction into a new utility function
