+++ gazette-parse/utils/parsePdfAsImageAndExtractText.ts
